Rename pages components to PascalCase
