+++ src/pages/BeaconRecords.tsx
